Update ehcache 2.7.4 -> 3.8.1 and activeJDBC 2.2 -> 2.3
The current verison of ehcache will cause memory leaks when queries are sent to the
gateway as ehcache will never clear statistics objects it creates about the queries.

There is no way to fix this in the current settings because activeJDBC locks `statistis=true`

This solution will update ehcache to use a new version of ehcache and activeJDBC that
does not have this issue. This also requires a change in the ehcache configuration file.