Support proxying /oauth2 to enable OAuth authentication (#190)
Hi folks,

Thank you for open-sourcing the presto-gateway, it's a very nifty piece
of software!

We're currently using it with Trino backends and [OAuth2
authentication](https://trino.io/docs/current/security/oauth2.html),
which works very well on its own but breaks when clusters are put behind
the presto-gateway, as OAuth2 does an HTTP callback to Trino under
`/oauth2`. This returns a 404 as the presto-gateway doesn't know about
this path and treats is as an error.

This small PR whitelists paths starting with `/oauth2` and instructs
presto-gateway to reverse-proxy it to a backend, enabling authenticating
via OAuth2.

Let me know if there's anything else I can do to make this PR more
likely to land, as we're happy using upstream rather than a fork. :)

Thank you!